MediaWiki: Massively reduce the number of lines of code. However, this breaks the API from MediaWiki 0.01: Category.getFullName() is now Category.getFullPageName() for consistency; all data classes return Long (or null) instead of long (or -1) for getPageID() and throw IOException in constructors.
Make AbstractContinuableQueryIterator contain more of the boilerplate code. Make it more robust in the face of exceptions, as well as more efficient.

Make createApiGetUrl(Map) accept and drop arguments that are mapped to null, as well as honour setMaxLag(). Drop the addition of params.put("maxlag", getMaxLag()) everywhere else.

Clarify the order of 'start' and 'end' parameters throughout the code by making them 'earliest' and 'latest' for time-based ranges or 'first' and 'last' for lexicographical ranges.