MediaWiki, WikiShell: Add support for title normalisation as well as parsing wikitext.
WikiShell: Add support for expanding {{PAGENAME}}, {{PAGENAMEE}} and a bunch of other page-related magic words in the 'download' and 'upload' commands' local file names, as well as the target name of a 'move' command (for example, for {{FULLPAGENAME (film)}}). Can be preset into commands made with the 'make' command. Also a fix for auto-buffering after newlines.

MediaWiki: Documentation tag fix.